add the Cirrus CI system
This automates build testing in the following platforms:
  - Alpine Linux (musl Linux)
  - Debian Linux (glibc Linux)
  - FreeBSD
  - MacOS

Multiple branches of Debian are used to cover toolchains ranging from
ancient to bleeding edge.
It would be very easy to add testing to this in the future, but
currently scrot has no tests.

If this is merged in, @eribertomota will have to add the
[Cirrus CI](https://cirrus-ci.org/) application to his github account
to enable it.

Cirrus can be seen in action at this URL: https://cirrus-ci.com/github/guijan/scrot/add-cirrus