polish the txt2man manual
I've made changes to the txt2man manual.

The most innocuous ones are the minor English corrections and making the description like the README.md file in the source tree.
I've changed the line length to be at most 80 characters always, 80 being the "standard" amount of columns in a terminal.

The options are sorted alphabetically and case insensitively by their short forms. The upper case short options don't all come first.
The synopsis is now a complete summary of all the flags.
Some flags which didn't previously list an argument but took one now have the fact that they take arguments clearly specified. I'm not 100% sure I got all of them though.
The manual now uses exclusively the short options while explaining its flags, to encourage their use.

This isn't all but that's the gist of it.