add libbsd dependency
This patch does a few things all at once:
- add libbsd to configure.ac
By default, the configure script picks up libbsd if it is necessary. A
"--without-libbsd" ./configure option was added to allow package maintainers
to test whether the package is needed on their OS without poking into the
source or removing libbsd from their local installation.

- add the libbsd effort to TODO.md
I also deleted the plaintext TODO, which was nothing but a single line joke.

- add libbsd package to Cirrus CI and Github Actions
Currently, the libbsd package is only installed on Debian and Ubuntu among our
tested systems because all the other systems provide all the currently used
functions.
Because they were on the same line, I also removed the giblib and libtool packages
from the Github Actions CI-scrot does not currently depend on either.

- add the first usage of each of err(), strlcat(), and strlcpy().
It turned out that the "err" identifier was in use in the code, so a
variable had to be renamed.

- add libbsd to the dependency list at README.md

- point to TODO.md from CONTRIBUTING.md