refactor options.c
This patch is a small refactor for options.c and its header.
A function had its name changed so a call in main.c had to be changed.

By using designated initializers, we can do away with one function and
abbreviate the code without being cryptic.
This also replaces a preprocessor use with an enum, and renames a function
to use a shorter name.