assume c89 functions exist
This changes configure.ac to assume c89 functions exist.
Checking for them is pedantic and scrot is not compatible with C89
or pre-ANSI C, scrot is written in the newer C99 standard.

The more obtuse and roundabout the code is the harder it is to
understand and maintain. Additionally, this will speed up the build
a little as autotools is fairly slow.