remove obsolete autotools macros
This patch:
- removes the the still hardcoded include directories in Makefile.am.
- updates the AC_CONFIGURE_HEADER macro to AC_CONFIGURE_HEADERS
Our CI says it's obsolete, I can't find the exact autotools announcement about
this, but here's someone saying it's obsolete already in 2006:
https://savannah.gnu.org/support/?105403
- removes the obsolete AM_PROG_CC_STDC macro
Oddly, we were using this obsolete macro as well as its replacement
(AC_PROG_CC) at the same time.
look for AM_PROG_CC_STDC at https://autotools.io/forwardporting/automake.html
- adds the AC_PROG_CC_C99 macro
It sets the compiler to C99 mode if that isn't the default. This has the side
effect of making the build work with ancient GCC versions in which C99 is
supported and C89 is the default, e.g. GCC 4.2.
- removes the AC_PATH_X macro
We get X compiler flags from pkg-config now. Additionally, the X flags provided
by this macro weren't actually used anyway.
- removes checks under the "checks for typedefs, structures, ..." comment
All of those except the stdbool one check for C89 features.
- removes the AC_FUNC_MALLOC macro
This macro is a workaround for malloc implementations that don't behave like
the GNU one when a malloc(0) call is made (allocating 0 bytes is
implementation-defined behavior in C). It's trivial to write correct C where
this isn't a problem, and scrot has few malloc/calloc calls and already
seems to do it.