bring the logo up to speed and use it
This patch does several things all at once:
- delete scrot-128x128.png

It's not used in the project and we're not a GUI program, so what is it
for?

- move the layer containing the "camera flash" effect in the .xcf logo
to the top layer

It's a bit weird if the camera's flash is behind the camera.

- move the camera flash one pixel to the left

This seems to center it better.

- output the lightly modified logo to doc/scrot.png and add the logo to
the README

I noticed that this project https://github.com/kohler/gifsicle has their
fancy logo at the top of the readme and I thought it was pretty cool. I
think we should do the same.