make functions static
This makes functions that aren't used across translation units static.
It also includes some style fixes.

Some but not all functions were moved to make the function declaration
and definition order closer to call order, however there are just too
many for me to do this to all of them right now.

Some lines that were over 80 columns were broken up.

Some declarations of functions belonging to other source files and their
headers were removed and the right header was included.

The '*' symbol in struct, variable and function declarations was moved
to be adjacent to the identifier and not the type, as per the WebKit
style for C programs.

An explicit cast was removed which seems to have revealed a compiler
warning.