Small CI revamp
- Add Cygwin to the CI

This is mainly to test portability further.
Cygwin uses the newlib libc on top of Windows. The newlib libc
implements a lot of BSD functions. It's a weird platform so maybe it'll
catch a bug in the future.

- Add distcheck CI target

I noticed that in PR #175 @daltomi noticed the PR broke
`make distcheck`, this could have been caught by the CI, now it will.
I also noticed that the Github Actions CI was doing almost all of
https://www.gnu.org/software/automake/manual/html_node/Checking-the-Distribution.html
by hand, so the lines were deleted as `make distcheck` now fills
their role.

- Bring the Cirrus CI in line with Github Actions

The Cirrus actions I wrote didn't run `scrot -v` or do the equivalent of
`make distcheck`, now they do.
I had to add the gzip package to the Alpine Linux VM as distcheck
depends on it.

- Update MacOS version in Cirrus CI

- Update actions/checkout in Github Actions CI