remove i386 ifdef
I finally understood what that #ifdef was: it was a micro optimization.
The #ifdef avoids a copy if xcim->pixels happens to be uint32, this is
the case on 32-bit Unix. That's the last #ifdef in scrot being removed.
I want to do this because it's easier to make sure scrot works on
various platforms if they're all running the same code.

This also adds a dependency on my libobsd library under OS X because
scrot now uses reallocarray() which isn't provided by OS X. My library
isn't packaged by Brew.