Get the name of the window with the field res_class
While the 'res_name' field should return the name of the application,
this doesn't happen correctly for all applications.

For example the application 'MPV' returns 'gl' as name.

The 'res_class' field will be used as the formal name.
Note that this is a resource database name so the first letter is
capitalized.