fix shift error
the PIXEL_ARGB() macro would do an overwide shift if `int` is 16bits.
and in case of 32bit ints the result of the shift `a << 24` can
overflow, which is undefined behavior.

test case, (compile with UBSan):

	$ ./configure CFLAGS="-fsanitize=undefined -g"
	$ make
	$ ./src/scrot -shide,some.png -l "opacity=254" -o /tmp/whatevr.png
	scrot_selection.c:426:25: runtime error: left shift of 156 by 24 places cannot be represented in type 'int'

this patch also removes the PIXEL_* macros entirely, they weren't needed
because only the alpha changed, r,g,b was unchanged.