fix libbsd feature test confusion
Users expressed confusion when using the old "--without-libbsd"
configure flag. Some assumed that it compiles scrot without libbsd, but
it only made the build script error if libbsd is a dependency. This is
useful to package maintainers: it gives them a yes/no answer on whether
the library is a dependency on their system.

The old configure flag has been replaced by a more verbose
"--enable-libbsd-feature-test".