Indicate that the value is not an integer
Until version v1.7 it would silently convert a floating point number
to a zero(0). Which is an error since it allowed the user to believe
that the value given to the option was correct, when it is not.
Since version v1.8, the entered value is indicated with an
error 'is not an integer'

Fix: #213