Pass SA_NODEFER to sigaction
Picture a situation in which a signal is delivered 3 times practically
instantly:
The 1st signal is caught and the handler starts. The default behavior of
signal handlers means the caught signal is blocked while the signal
handler runs. The 2nd signal is blocked. When the 3rd signal is
received, because the same signal is already pending, POSIX says the 3rd
signal can be discarded.

Passing this one flag fixes the issue.