scrotDoDelay: output countdown to stderr
otherwise, the countdown won't be visible when user is outputting the
image to stdout. for example commands like the following will end up
with a "corrupted" image due to the countdown being slipping in there.

	scrot -d3 -c - | xclip -in -selection clipboard -target image/png

by outputting the countdown to stderr, the above command will work
properly.

for reference, `maim` also outputs it's countdown to stderr, not stdout.