fix multiple bugs in scrot's -C option
Previously, scrot arbitrarily and silently confined the 2nd string of an
X11 WM_CLASS (the res_class member of Xlib's XClassHint structure) to a
length of 79 characters.

Additionally, scrot only compared those first 79 characters when
matching windows. Fixing this bit changes scrot's behavior, but it's
undocumented behavior, not useful, and window class names that long are
expected to be few.

Also, scrot's handling of the -C option suffered from a memory leak. It
would leak class names as many times as the -C option was repeated.

This also comes with a few improvements that aren't necessarily bugs:
- A copy of the window class name requested by the user was
  unnecessarily created with strndup(), this change removes the
  unnecessary copying.
- scrot used to ignore -C if its argument was the null string. If the -C
  option with a null string argument was the only -C option, it would
  match any window class name (which is the default behavior). If any -C
  option preceded it, it wouldn't replace the previous -C option. Now it
  tries to match a null window class name.