Deprecate -S
According to https://docs.enlightenment.org/api/imlib2/html/ section
Imlib 2 Dynamic Filters, an imlib2 script can contain an "[] operator"
which pops a pointer from the varargs and reads it. This means the
-S functionality suffers from an unfixable bug: the user can put an
arbitrary amount of "[]" operators in the script, and there is no
way in C to pass the right amount of varargs to match this.

Additionally, varargs are typically implemented by passing values on the
stack and popping them. This means the stack pointer can be decremented
from argv. Even more worrying, argv typically resides on the stack, so
barring OS mitigations, scrot's caller can decrement the stack until it
points to attacker-controlled memory. This is a severe security
vulnerability.

Mark -S for removal in a future release.