imPrintf: avoid going out of bound
when the last char is `$` or `\`, the `c` pointer would end up advancing
past the string length.

additionally use the return value of strftime for determining the string
length instead of checking for '\0'.

also replaces hardcoded length in strftime call to sizeof(strf).
previously it was one less than sizeof(strf), but there was no real
reason for that. strftime will return 0 if the resulting string
(including the nul-byte) was too big.

Fixes: https://github.com/resurrecting-open-source-projects/scrot/issues/252