print excess arguments one line at a time
pretty much the only practical reason this would occur is if someone is
unfamiliar with shell and tries something like `scrot -F my file name`.
in which case, he'll get the following warning:

	scrot: ignoring extraneous non-option arguments: file name

which might not make it obvious as to what's going on. just printing
each extra arg on it's own line makes it obvious that the space is
causing the arguments to be split. the code also shorter this way as an
added bonus.