improve timing in `scrot --delay`
Upgrade scrotSleep() to scrotWaitUntil() which sleeps until a given time
on the clock rather than sleeping for a given period of time. In all
cases, this improves the timing by compensating for the time passed
between entering main() and taking the screenshot. When used with
--count, it fixes the drift caused by calling nanosleep() in a
loop.

Also remove an incorrect comment and rename the delay_selection variable
to delaySelection to match our style.