remove some useless assert()s
All of the removed assert()s are checking whether pointers that are soon
to be dereferenced by scrot are NULL. There are many more pointers in
the source code that don't get assert()ed and any OS these days will
have NULL == 0 and an unmapped 0-page that causes the program to
segfault on read/write, the exception would be MMU-less systems, but we
don't expect people to develop on those.

The only exception is scrotGetWindowName(), but the assert() redundantly
checks if the function's argument is NULL, when its only caller already
does that without assert().