scrotGetGeometry: sleep explicitly
this patch simply removes the fluff and indirection and shows what's
actually going on already. since we never subscribed to visibility
event, we're never going to receive it anyways and will end up sleeping
the entire duration.

also reduce the sleep from ~300ms to ~160ms, which should be enough for
local X server. (note: it's fine to call nanosleep in loop, it's just a
fuzzy sleep, doesn't have to be exact).

ref: https://github.com/resurrecting-open-source-projects/scrot/issues/274