ci: check for bare build as well
with the previous two commits, dependency on autoconf is pretty much
gone (aside from the libbsd check). so add a simple, no autoconf build
to the ci to ensure we keep this intact.

and since musl (alpine-latest) doesn't have sys/queue.h, require
libbsd-overlay in addition to ./deps.pc there.

this "bare build" is pretty much undocumented for now, as it is
experimental. if it goes well, we can document it in the README.