README: depend on an interface, not an implementation
pkgconf has reimplemented pkg-config and replaced it in the wild. For
instance, in Debian: https://packages.debian.org/sid/pkg-config
The original pkg-config implementation by freedesktop is abandoned, the
last release was in 2017-03-20:
https://pkgconfig.freedesktop.org/releases/

We don't care what implementation is in use, just describe the interface
and provide an example. Use more verbose language because both the
interface and an implementation have the same name which could cause
confusion.

We already did this for the X server, but also change the language used
there to make it consistent.