scrotGrabMousePointer: avoid malloc and fixup in place
initially this path wasn't chosen because we suspected this might
violate C's "strict-aliasing" rule. but a more careful look shows that
this is not the case.

xcim->pixels (due to being dynamically allocated memory) does not have
any declared type, and so it's effective type is determined by the last
write, in our case `unsigned long`.

	pixels[i] = xcim->pixels[i];

in all iteration of the loop, `xcim->pixels[i]` at the time of the read,
has the effective type of `unsigned long` which is what we are reading
it as, so there's no issue.

the `pixels[i]` write changes the written-to region's effective type to
`uint32_t` but that value is never read until the `imlib_create_image_using_data`
call later, which reads it as `uint32_t` - matching it's effective type.

as such, there's no real strict-aliasing violation here and we can
simply fixup the array in place. this reduces a potential runtime error
and (in theory) uses less memory.

ref: https://port70.net/~nsz/c/c99/n1256.html#6.5p6
ref: https://github.com/resurrecting-open-source-projects/scrot/pull/193#discussion_r962428111