configure.ac: remove unsound/misleading flag
this avoids misleading people due to the issues described in #315.

and packagers can already check the output of the configure script or
config.log to see if libbsd was checked for or not.

additionally, document which bsd extension we require in the README so
that maintainers can make more informed decision (e.g using more
lightweight `sys/queue.h` in musl instead of heavyweight libbsd).

ref: https://github.com/resurrecting-open-source-projects/scrot/issues/307
Fixes: https://github.com/resurrecting-open-source-projects/scrot/issues/315