imPrintf: use dynamic buffer for strftime
this avoids unnecessary limit on the filename length (at the cost of
creating new failure points due to dynamic allocations).

the biggest annoyance, and the reason why an additional dynamically
allocated buffer was necessary, is the fact that strftime() doesn't have
a way to distinguish between valid 0 return vs 0 return due to small
buffer.

prepending a useless '.' in the format buffer and skipping over it later
on solves this issue since it can no longer "validly" return 0.

Fixes: https://github.com/resurrecting-open-source-projects/scrot/issues/223