fix: buffer overrun due to bad x,y
imlib_create_image_from_ximage() expects x and y relative to the image
itself and has no relation with the x,y position of the window.

which means we should pass 0 in both of them. otherwise, it'll try to
read a width and height that's bigger than the image actually has.

buffer overflow was caught via AddressSanitizer.