fix weird optional argument parsing
currently if scrot is invoked like `scrot --stack=weird=v` then it
treats as if the argument to --stack was `v` and skips over "weird="
entirely.

the code here seems to be trying to get rid of the '=' but there's no
need for it, getopt_long() will remove the '=' already.